Move all test files up to test root dir
