+++ browser/test/AutoClosingPairsTests.ts
