+++ browser/test/CompletionUtilityTests.ts
