+++ browser/test/ExplorerSelectorsTests.ts
