+++ browser/test/InputManagerTests.ts
