+++ browser/test/LanguageManagerTests.ts
