+++ browser/test/SpanTests.ts
