+++ browser/test/TypingPredictionManagerTests.ts
