+++ browser/test/WindowSplitTests.ts
